fix issues cause by deprecated numpy float

+++ pyiron_atomistics/lammps/units.py
@@ -155,9 +155,9 @@ for key, val in _conversion_dict.items():
     for v in val:
         # Everything except dimensionless integer quantities are stored as floats
         if key in ["dimensionless_integer_quantity"]:
-            dtype_dict[v] = np.int
+            dtype_dict[v] = int
         else:
-            dtype_dict[v] = np.float
+            dtype_dict[v] = np.float64
 
 
 class UnitConverter:
